Fix Windows support
