+++ Lib/_pyrepl/simple_interact.py
@@ -25,14 +25,13 @@ allowing multiline input and multiline history entries.
 
 from __future__ import annotations
 
-import _colorize  # type: ignore[import-not-found]
 import _sitebuiltins
 import linecache
 import sys
 import code
-import ast
 from types import ModuleType
 
+from .console import InteractiveColoredConsole
 from .readline import _get_reader, multiline_input
 from .unix_console import _error
 
@@ -64,48 +63,6 @@ REPL_COMMANDS = {
     "help": "help",
     "clear": "clear_screen",
 }
-
-class InteractiveColoredConsole(code.InteractiveConsole):
-    def __init__(
-        self,
-        locals: dict[str, object] | None = None,
-        filename: str = "<console>",
-        *,
-        local_exit: bool = False,
-    ) -> None:
-        super().__init__(locals=locals, filename=filename, local_exit=local_exit)  # type: ignore[call-arg]
-        self.can_colorize = _colorize.can_colorize()
-
-    def showsyntaxerror(self, filename=None):
-        super().showsyntaxerror(colorize=self.can_colorize)
-
-    def showtraceback(self):
-        super().showtraceback(colorize=self.can_colorize)
-
-    def runsource(self, source, filename="<input>", symbol="single"):
-        try:
-            tree = ast.parse(source)
-        except (OverflowError, SyntaxError, ValueError):
-            self.showsyntaxerror(filename)
-            return False
-        if tree.body:
-            *_, last_stmt = tree.body
-        for stmt in tree.body:
-            wrapper = ast.Interactive if stmt is last_stmt else ast.Module
-            the_symbol = symbol if stmt is last_stmt else "exec"
-            item = wrapper([stmt])
-            try:
-                code = self.compile.compiler(item, filename, the_symbol)
-            except (OverflowError, ValueError):
-                    self.showsyntaxerror(filename)
-                    return False
-
-            if code is None:
-                return True
-
-            self.runcode(code)
-        return False
-
 
 def run_multiline_interactive_console(
     mainmodule: ModuleType | None = None,
